requirements.txt
streamlit==1.35.0
pydantic==2.7.1
fpdf==1.7.2
educhain==0.1.18
google-generativeai==0.5.4
langchain-google-genai==0.1.7
langdetect==1.0.9
SpeechRecognition==3.10.1
pydub==0.25.1
python-dotenv==1.0.1
googletrans==4.0.0rc1
requests==2.31.0